add ci, start working on linter conformation
