--- README.md
@@ -1,6 +1,0 @@
-# AwkwardQL
-
-A SQL-like language for doing set operations on data in Awkward Arrays. 
-This is derived from and inspired by the [demo](https://github.com/jpivarski/PartiQL) by Jim Pivarski (PartiQL) and targets awkward 1.0.
-
-We will start out focusing on particle physics uses cases and see how far it goes from there. :-)
