+++ README.rst
@@ -1,0 +1,41 @@
+AwkwardQL - SQL-like language for awkward arrays
+================================================
+
+.. inclusion-marker-1-do-not-remove
+
+A SQL-like language for doing set operations on data in Awkward Arrays.
+
+.. inclusion-marker-1-5-do-not-remove
+
+This is derived from and inspired by the demo by Jim Pivarski (PartiQL) and targets awkward 1.0.
+
+We will start out focusing on particle physics uses cases and see how far it goes from there. :-)
+
+.. inclusion-marker-2-do-not-remove
+
+Installation
+============
+
+Install coffea like any other Python package:
+
+.. code-block:: bash
+
+    pip install awkwardql
+
+or similar (use ``sudo``, ``--user``, ``virtualenv``, or pip-in-conda if you wish).
+
+Strict dependencies
+===================
+
+- `Python <http://docs.python-guide.org/en/latest/starting/installation/>`__ (3.6+)
+
+The following are installed automatically when you install coffea with pip:
+
+- `awkward1 <https://github.com/scikit-hep/awkward-array>`__ to manipulate complex-structured columnar data, such as jagged arrays;
+- `numba <https://numba.pydata.org/>`__ just-in-time compilation of python functions;
+
+.. inclusion-marker-3-do-not-remove
+
+Documentation
+=============
+Not yet developed, but you should check out the notebooks in this repository!
