+++ awkwardql/__init__.py
@@ -1,0 +1,3 @@
+from . import version
+
+__version__ = version.version
