+++ tests/test_import.py
@@ -1,0 +1,4 @@
+import pytest
+
+def test_import():
+    import awkwardql as awkwardql
